Refactored the languages control logic
Added a checkAvailableLanguages flag to the settings. If this is false
no attempt is made to pull the languages.json file and you may get some
404 errors in the dev console. The behaviour in this case will be the
same as it was in 1.0.9. If true, the languages file is pulled and the
codes are checked against it before a download is attempted.