Issue #25 Added async option
If you supply the flag async on the settings and set it to true, all
ajax calls are executed asynchronously and the supplied callback is
called after the languaged files have all been downloaded and parsed. If
you leave the flag off, or set it to false, the behaviour is as before:
all the files are parsed synchronously and the callback is called at the
end of the process.